+++ lib/Widgets/floater.dart
@@ -15,12 +15,11 @@ class Floater extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    Brightness brightness = MediaQuery.of(context).platformBrightness;
     return Container(
       margin: margin ?? const EdgeInsets.all(10),
       child: Material(
         borderRadius: BorderRadius.circular(10),
-        color: backgroundColor ?? (brightness == Brightness.dark ? Colors.black : Colors.white),
+        color: backgroundColor ?? Colors.white,
         elevation: elevation ?? 10,
         child: child,
       ),